requirements.txt
boto3==1.28.54
botocore==1.31.54
pandas==2.1.1
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.16
openpyxl==3.1.2
numpy==1.26.4
sqlalchemy==2.0.31